created useFetchUsers hook, split code
