+++ src/hooks/index.js
@@ -1,1 +1,2 @@
 export { default as useTitle } from "./useTitle";
+export { default as useFetchUsers } from "./useFetchUsers";
