+++ src/pages/SearchPage/SearchPage.styled.jsx
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const UserListContainer = styled.div`
+  height: 100%;
+  overflow-y: auto;
+  overflow-x: hidden;
+  scrollbar-width: thin;
+`;
+
+const ListOptions = styled.div`
+  height: 40px;
+`;
+
+export { UserListContainer, ListOptions };
